Move onSample function back to deviceactions
